+++ __tests__app__/dev/src/index.js
@@ -34,9 +34,6 @@ const app = new Server({
 		path.resolve('./node_modules/bootstrap/dist/css/bootstrap.css'),
 	], */
 	rsConfig: path.resolve(__dirname, '../rsconfig.json'),
-	cookies: {
-		secure: false,
-	},
 	// createMissingComponents: true,
 	// autoprefixer: { grid: 'autoplace' },
 	// cookieSecret: 'dev-secret',